backend/requirements.txt
fastapi
uvicorn[standard]
pymongo
motor
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
pydantic
pydantic-settings
email-validator
aiohttp
requests
python-multipart
bcrypt
apscheduler
pytz
